requirements.txt
accelerate==0.27.2
acres==0.3.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.3.0
Authlib==1.5.2
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.4
boto3==1.38.27
botocore==1.38.27
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
coloredlogs==15.0.1
comm==0.2.2
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.0
cryptography==44.0.2
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
emoji==2.14.1
etelemetry==0.3.1
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.12
filelock==3.18.0
filetype==1.2.0
fitz==0.0.1.dev2
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2025.3.2
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.18.1
isodate==0.6.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
langchain==0.1.13
langchain-community==0.0.32
langchain-core==0.3.72
langchain-openai==0.3.28
langchain-text-splitters==0.0.2
langdetect==1.0.9
langsmith==0.4.10
looseversion==1.3.0
lxml==5.4.0
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.9.4
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.4.3
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.2.1
nibabel==5.3.2
nipype==1.10.0
nltk==3.9.1
numpy==1.26.4
olefile==0.47
onnx==1.17.0
onnxruntime==1.19.2
openai==1.98.0
opencv-python==4.11.0.86
orjson==3.10.18
packaging==23.2
pandas==2.2.1
parso==0.8.4
passlib==1.7.4
pathlib==1.0.1
pdf2image==1.17.0
pdfminer.six==20250416
pexpect==4.9.0
pi_heif==0.22.0
pillow==11.2.1
platformdirs==4.3.7
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==6.30.2
prov==2.0.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.28
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydot==3.0.4
Pygments==2.19.1
PyMuPDF==1.25.5
pyparsing==3.2.3
pypdf==5.4.0
PyPDF2==3.0.1
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-iso639==2025.2.18
python-jose==3.4.0
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
pyzmq==26.4.0
RapidFuzz==3.13.0
rdflib==6.3.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
s3transfer==0.13.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==4.1.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
stack-data==0.6.3
starlette==0.46.2
sympy==1.13.1
tenacity==8.5.0
threadpoolctl==3.6.0
tiktoken==0.9.0
timm==1.0.15
tokenizers==0.21.1
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
transformers==4.47.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
unstructured==0.17.2
unstructured-client==0.34.0
unstructured-inference==0.8.10
urllib3==1.26.20
uvicorn==0.34.2
validators==0.34.0
wcwidth==0.2.13
weaviate-client==3.25.0
webencodings==0.5.1
wrapt==1.17.2
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0
